requirements.txt
fastapi
uvicorn
python-telegram-bot[ext]
sqlalchemy
yt-dlp
python-dotenv
httpx
